agriculture-federated-learning/archive/canopy_height/requirements.txt
matplotlib==3.8.4
numpy==1.26.4
Pillow==10.3.0
tifffile==2024.4.18
